lntest: move multi-hop root test case and common utils to own file
